Added control frames for multi-file transmission
Control frames are the mechanism used to delimit file boundries
during multi-file transmission. The current strategy is to send
a preamble frame before any data is read and an EOT frame after
we have transmitted everything. Tx comes with the option to send
multiple redundant control frames so that we can ensure that the
receiver sees and processes the control frame. The receiver
recognizes that the captured packet is a control frame by checking
the size of the payload (It should match the control frame size constant)
and then verify that the majority of the payload data matches a
control data value.