Refactored directory sniffer into a new module
Adds the dirwatch library module that provides a much more focused
API for monitoring a directory for new files. DirWatch is built
on top of Inotify, but it handles dealing with Inotify's cumbersome
API and quirks so you don't have to. Currently, dirwatch only
monitors for the Create and Close event combo, but can be easily
modified to support a number of different events.