Asserts now uses the logger for output
The reason behind this change is to ensure that the logs have data
about fatal runtime conditions. Assert was just piping everything to
stderr, but if we are running the tx/rx programs in a headless and
remote environment we won't have access to stderr. in the background
in a headless environment then we won't have access to stderr.