Fix number of bugs in decoder
I never got a change to actually review PR #12 before it was merged.
Now I'm having a chance to work through it I found a number of bugs.

- Searching for a valid OTI header didn't actually do anything since
  the code just used the first OTI in the message after the search.

- Decode was always reporting success even on failure because it
  was capturing error codes into an unsigned value and checking
  that the value was > 0

Worst of all, the RS decoder is not actually working! This was
my primary concern when I handed off the FEC stuff, I didn't think
the RS decoder was actually correcting for any bit errors. This
bug I did not fix yet, and will make a new issue for.