Remove FCS handling in transmitter
Previously we would allocate an extra four bytes into each frame
and direct the driver to use that space for the FCS. Further
testing revealed that the driver would actually elide the last
four bytes of every packet. We now remove FCS handling which
results in the driver attaching four extra bytes to the end of
our frames.